backend/pyproject.toml
[project]
name = "yt-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.76.2",
    "python-dotenv>=1.1.0",
    "pytube>=15.0.0",
    "uvicorn>=0.34.2",
    "youtube-transcript-api>=1.0.3",
]
